style(Event): rename a event
