+++ pvzclass/Events/Events.h
@@ -5,7 +5,7 @@
 #include "CoinRemoveEvent.h"
 #include "DialogButtonDepressEvent.h"
 #include "DialogDrawEvent.h"
-#include "DrawUITopEvent.h"
+#include "DrawUIEvents.hpp"
 #include "DrawObjectEvents.hpp"
 #include "DrawPlantReanimEvent.h"
 #include "DrawZombieReanimEvent.h"
